requirements.txt
streamlit==1.10.0
pandas==1.3.5
numpy==1.21.5
audio-recorder-streamlit
python-dotenv
openai